+++ src/pages/Pools/components/WithdrawLiquidityModal/WithdrawLiquidityModal.module.scss
@@ -288,6 +288,7 @@
         display: flex;
         justify-content: space-between;
         gap: 8px;
+        margin-top: 10px;
 
         .item {
           display: flex;
